backend/requirements.txt
# Core Django and API framework
Django==4.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Cache and Redis
redis==5.0.1
django-redis==5.4.0

# External API integrations
google-generativeai==0.3.2
openai==1.3.7
requests==2.31.0

# CORS and security
django-cors-headers==4.3.1
django-environ==0.11.2

# File uploads and storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0

# Filtering and utilities
django-filter==23.4
django-extensions==3.2.3

# Async tasks (optional but recommended)
celery==5.3.4
kombu==5.3.4

# Monitoring and logging
sentry-sdk==1.38.0

# Development tools
django-debug-toolbar==4.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Documentation
drf-yasg==1.21.7

# Deployment
python-decouple==3.8